pyproject.toml
[project]
name = "machine-translation"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.10,<3.15"
dependencies = [
    "torch (>=2.10.0,<3.0.0)",
    "sentencepiece (>=0.2.1,<0.3.0)",
    "pandas==2.3.3",
    "tqdm (>=4.67.3,<5.0.0)",
    "sacrebleu (>=2.6.0,<3.0.0)",
    "pandas-stubs~=2.3.3",
    "pyyaml>=6.0.3",
    "streamlit>=1.55.0",
    "ipykernel>=7.2.0",
]
